+++ changelog.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- process CLI args with getopt
+
 v1.3.1 - 2026-05-07
 - Added macros to call for cleanup and errors. This cleans up the code
   a bit and makes it more standarized (less code duplication and